// src/api/__generated__/mining/model/actorDTOStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

/**
 * 派遣状态: MINING/REST
 */
export type ActorDTOStatus =
  (typeof ActorDTOStatus)[keyof typeof ActorDTOStatus];

export const ActorDTOStatus = {
  REST: 'REST',
  MINING: 'MINING',
} as const;